+++ src/main/webapp/resources/js/pages/admin/routes.js
@@ -1,0 +1,14 @@
+/*
+ * This file contains the route names for the admin panel
+ */
+
+export const ADMIN = {
+  ADMIN: "admin",
+  USERS: "users",
+  USERGROUPS: "groups",
+  CLIENTS: "clients",
+  REMOTEAPI: "remote_api",
+  SEQUENCINGRUNS: "sequencing_runs",
+  NCBIEXPORTS: "ncbi_exports",
+  ANNOUNCEMENTS: "announcements"
+};